Rename files to camelcase
